apps/worker/requirements.txt
boto3
coverage
python-dotenv
requests
pydantic
pydantic[email]